refactor: fixing up
